+++ rolldown.config.ts
@@ -2,6 +2,6 @@ import { dts } from 'rolldown-plugin-dts'
 
 export default {
   input: './src/index.ts',
-  plugins: [dts()],
-  output: [{ dir: 'dist', format: 'es', sourcemap: true }]
+  output: [{ dir: 'dist', format: 'es', sourcemap: true }],
+  external: ['pdfjs-dist']
 }